pyproject.toml
[project]
name = "log-listener-proxy"
version = "0.1.0"
description = "Микросервис для проксирования stdout/stderr от PySpark-приложений к клиентам через WebSocket"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",        # веб-фреймворк с поддержкой WebSocket
    "uvicorn[standard]>=0.27.0", # ASGI-сервер для запуска FastAPI
    "websockets>=12.0"        # WebSocket-библиотека для FastAPI
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",           # тестовый фреймворк
    "pytest-asyncio>=0.23.0",  # асинхронные тесты
    "requests>=2.28.0",        # HTTP-клиент для тестов
    "websocket-client>=1.6.0", # WebSocket-клиент для тестов (совместим с Python 3.7+)
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["log_listener_proxy"]
